Fix namespace of TrashItemHandler
