+++ packages/page/src/Infrastructure/Sulu/Trash/PageTrashItemHandler.php
@@ -11,14 +11,8 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Article\Trash;
+namespace Sulu\Page\Infrastructure\Sulu\Trash;
 
-use Sulu\Article\Application\Mapper\ArticleMapperInterface;
-use Sulu\Article\Domain\Model\ArticleDimensionContent;
-use Sulu\Article\Domain\Model\ArticleDimensionContentInterface;
-use Sulu\Article\Domain\Model\ArticleInterface;
-use Sulu\Article\Domain\Repository\ArticleRepositoryInterface;
-use Sulu\Article\Infrastructure\Sulu\Admin\ArticleAdmin;
 use Sulu\Bundle\TrashBundle\Application\RestoreConfigurationProvider\RestoreConfiguration;
 use Sulu\Bundle\TrashBundle\Application\RestoreConfigurationProvider\RestoreConfigurationProviderInterface;
 use Sulu\Bundle\TrashBundle\Application\TrashItemHandler\RestoreTrashItemHandlerInterface;
@@ -29,51 +23,59 @@ use Sulu\Content\Application\ContentMerger\ContentMergerInterface;
 use Sulu\Content\Application\ContentNormalizer\ContentNormalizerInterface;
 use Sulu\Content\Domain\Model\DimensionContentCollection;
 use Sulu\Content\Domain\Model\DimensionContentInterface;
+use Sulu\Page\Application\Mapper\PageMapperInterface;
+use Sulu\Page\Domain\Model\PageDimensionContent;
+use Sulu\Page\Domain\Model\PageDimensionContentInterface;
+use Sulu\Page\Domain\Model\PageInterface;
+use Sulu\Page\Domain\Repository\PageRepositoryInterface;
+use Sulu\Page\Infrastructure\Sulu\Admin\PageAdmin;
 use Webmozart\Assert\Assert;
 
 /**
  * @internal
  */
-final class ArticleTrashItemHandler implements
+final class PageTrashItemHandler implements
     StoreTrashItemHandlerInterface,
     RestoreTrashItemHandlerInterface,
     RestoreConfigurationProviderInterface
 {
     /**
-     * @param iterable<ArticleMapperInterface> $articleMappers
+     * @param iterable<PageMapperInterface> $pageMappers
      */
     public function __construct(
         private TrashItemRepositoryInterface $trashItemRepository,
-        private ArticleRepositoryInterface $articleRepository,
+        private PageRepositoryInterface $pageRepository,
         private ContentNormalizerInterface $contentNormalizer,
         private ContentMergerInterface $contentMerger,
-        private iterable $articleMappers,
+        private iterable $pageMappers,
     ) {
     }
 
     public static function getResourceKey(): string
     {
-        return ArticleInterface::RESOURCE_KEY;
+        return PageInterface::RESOURCE_KEY;
     }
 
     public function store(object $resource, array $options = []): TrashItemInterface
     {
-        Assert::isInstanceOf($resource, ArticleInterface::class);
+        Assert::isInstanceOf($resource, PageInterface::class);
 
-        $article = $resource;
+        $page = $resource;
 
         $data = [
+            'parentUuid' => $page->getParent()?->getUuid(),
+            'webspaceKey' => $page->getWebspaceKey(),
             'dimensionContents' => [],
         ];
 
         $restoreType = $options['locales'] ?? null ? 'translation' : null;
 
         $titles = [];
-        /** @var array<string, ArticleDimensionContentInterface> $localizedDimensionContents */
+        /** @var array<string, PageDimensionContentInterface> $localizedDimensionContents */
         $localizedDimensionContents = [];
-        /** @var ArticleDimensionContentInterface|null $unlocalizedDimensionContent */
+        /** @var PageDimensionContentInterface|null $unlocalizedDimensionContent */
         $unlocalizedDimensionContent = null;
-        foreach ($article->getDimensionContents() as $dimensionContent) {
+        foreach ($page->getDimensionContents() as $dimensionContent) {
             if (
                 DimensionContentInterface::CURRENT_VERSION !== $dimensionContent->getVersion()
                 && DimensionContentInterface::STAGE_LIVE !== $dimensionContent->getStage()
@@ -89,20 +91,20 @@ final class ArticleTrashItemHandler implements
             $localizedDimensionContents[$dimensionContent->getLocale()] = $dimensionContent;
         }
 
-        Assert::notNull($unlocalizedDimensionContent, 'Expected to find an unlocalized dimension content for the article.');
-        Assert::notEmpty($localizedDimensionContents, 'Expected to find at least one localized dimension content for the article.');
+        Assert::notNull($unlocalizedDimensionContent, 'Expected to find an unlocalized dimension content for the page.');
+        Assert::notEmpty($localizedDimensionContents, 'Expected to find at least one localized dimension content for the page.');
 
         // sort dimensionContents after the availableLocales from the unlocalizedDimension
         $availableLocales = $unlocalizedDimensionContent->getAvailableLocales();
         Assert::isArray($availableLocales, 'Expected availableLocales to be an array');
-        /** @var array<string, ArticleDimensionContentInterface> $localizedDimensionContents */
+        /** @var array<string, PageDimensionContentInterface> $localizedDimensionContents */
         $localizedDimensionContents = \array_merge(
             \array_flip($availableLocales),
             $localizedDimensionContents
         );
 
         foreach ($localizedDimensionContents as $locale => $localizedDimensionContent) {
-            /** @var array<int, ArticleDimensionContent> $dimensionContents */
+            /** @var array<int, PageDimensionContent> $dimensionContents */
             $dimensionContents = [$unlocalizedDimensionContent, $localizedDimensionContent];
 
             $mergedDimensionContent = $this->contentMerger->merge(
@@ -113,7 +115,7 @@ final class ArticleTrashItemHandler implements
                         'stage' => DimensionContentInterface::STAGE_DRAFT,
                         'version' => DimensionContentInterface::CURRENT_VERSION,
                     ],
-                    ArticleDimensionContent::class
+                    PageDimensionContent::class
                 )
             );
 
@@ -128,48 +130,65 @@ final class ArticleTrashItemHandler implements
         }
 
         return $this->trashItemRepository->create(
-            ArticleInterface::RESOURCE_KEY,
-            $article->getUuid(),
+            PageInterface::RESOURCE_KEY,
+            $page->getUuid(),
             $titles,
             $data,
             $restoreType,
             $options,
-            ArticleAdmin::SECURITY_CONTEXT,
+            PageAdmin::getPageSecurityContext($page->getWebspaceKey()),
             null, // TODO add Security
-            $article->getUuid(),
+            $page->getUuid(),
         );
     }
 
     /**
-     * @param array{} $restoreFormData
+     * @param array{
+     *     parentId?: string,
+     * } $restoreFormData
      */
     public function restore(TrashItemInterface $trashItem, array $restoreFormData = []): object
     {
         $restoreData = $trashItem->getRestoreData();
-        $articleUuid = $trashItem->getResourceId();
+        $pageUuid = $trashItem->getResourceId();
 
-        $article = $this->articleRepository->createNew($articleUuid);
-        $this->articleRepository->add($article);
+        // Create the page
+        $page = $this->pageRepository->createNew($pageUuid);
+        $webspaceKey = $restoreData['webspaceKey'];
+        Assert::string($webspaceKey, 'Expected webspaceKey to be a string');
+        $page->setWebspaceKey($webspaceKey);
 
-        $dimensionContents = $restoreData['dimensionContents'] ?? [];
-        Assert::isArray($dimensionContents, 'Expected dimensionContents to be an array');
-        /** @var array<string, mixed> $dimensionContentData */
-        foreach ($dimensionContents as $dimensionContentData) {
-            unset($dimensionContentData['url']); // TODO old route is not removed on delete?
-            foreach ($this->articleMappers as $articleMapper) {
-                $articleMapper->mapArticleData($article, $dimensionContentData);
+        // Set parent if exists
+        $parentUuid = $restoreFormData['parentId'] ?? $restoreData['parentUuid'];
+        if ($parentUuid) {
+            Assert::string($parentUuid, 'Expected parentUuid to be a string');
+            $parent = $this->pageRepository->findOneBy(['uuid' => $parentUuid]);
+            if ($parent) {
+                $page->setParent($parent);
             }
         }
 
-        return $article;
+        $this->pageRepository->add($page);
+
+        $dimensionContents = $restoreData['dimensionContents'] ?? [];
+        Assert::isArray($dimensionContents, 'Expected dimensionContents to be an array');
+        foreach ($dimensionContents as $dimensionContentData) {
+            Assert::isArray($dimensionContentData, 'Expected dimensionContentData to be an array');
+            unset($dimensionContentData['url']); // TODO old route is not removed on delete?
+            foreach ($this->pageMappers as $pageMapper) {
+                $pageMapper->mapPageData($page, $dimensionContentData);
+            }
+        }
+
+        return $page;
     }
 
     public function getConfiguration(): RestoreConfiguration
     {
         return new RestoreConfiguration(
-            'restore_article',
-            ArticleAdmin::EDIT_TABS_VIEW,
-            ['id' => 'id'],
+            'restore_page',
+            PageAdmin::EDIT_FORM_VIEW,
+            ['id' => 'id', 'webspace' => 'webspace'],
             null, // TODO serialization group?
         );
     }
